Corregido el tipo de fibo

+++ Userland/SampleCodeModule/lib/fibonacci.c
@@ -3,20 +3,20 @@
 #include <usersyscalls.h>
 
 int fibonacci(){
-  int a = 0;
-  int b = 1;
-  int c;
+  uint64_t a = 0;
+  uint64_t b = 1;
+  uint64_t c;
   int count = 2;
-  _fprintf(STDOUT, "n %d: %d\n", 1, 0);
+  _fprintf(STDOUT, "n %d: %u\n", 1, 0);
   sys_wait(1);
-  _fprintf(STDOUT, "n %d: %d\n", 2, 1);
+  _fprintf(STDOUT, "n %d: %u\n", 2, 1);
   while(1){
     c = a + b;
     a = b;
     b = c;
     count++;
     sys_wait(1);
-    _fprintf(STDOUT,"n %d: %d\n",count ,c);
+    _fprintf(STDOUT,"n %d: %u\n",count ,c);
   }
   sys_exit();
   return c;
